Initial proof of concept for colorset translations
This is a proof of concept for #167 and #169, adds basic functionality
for translation the names of colors in colorsets while still keeping
them modular.  @techninja and I talked about the main issues with this
still, mainly including the namespaces in the colorset JSON file, and
the general functionality (or lack thereof) of translations in modes.