Colorset translations WORK!!!!!!
After a VERY busy week of me not having time to work on robopaint, I
finally got colorset translations DONE and WORKING (for the most part)!
I was only able to port the colorsets over to English, but this gives
our translation gurus a template to work off of. Anyways, here's a
feature list:
- Added i18n folders and en-US translation files for respective
colorsets
- Modified colorset definition files to move all text elements to i18n
files.
- Modified colorset parser to get text from i18n file and reference a
non-keyed color array.