Initial creation of 'print' mode translation file
The start of my work to get mode translation working, I changed the
print mode to actually have the respective translate text in a translate
file instead of in the code directly. List of changes:
- Static translate strings embedded in the HTML have been changed to
translate strings
- Dynamic translate strings (status messages) are translated and added
to an array at the beginning of the mode JS file. This array is
referenced instead of a translate string (keeping all major translate
strings in the same place)
- Added all strings to the existing print i18n mode file.

TODOs:
- Actually write the backend code to translate the strings
- Create base English translations from existing text in other modes
- Create mode translations for other languages we support